Refactor after resolved conflicts.
