+++ src/app/exam/ExamList.js
@@ -6,8 +6,8 @@ import {ItemListPage} from "../commons/ItemListPage/ItemListPage";
 import {CreateButton} from "../commons/buttons/CreateButton";
 import {displayDate} from "../../utils/utils";
 import {CreateExamModal} from "./modals/CreateExamModal";
-import {Examinees} from "./Examinees";
-import ExamProblems from "./ExamProblems";
+import {Examinees} from "./participants/Examinees";
+import ExamProblems from "./problems/ExamProblems";
 import {ExamOptions} from "./options/ExamOptions";
 import './ExamList.css';
 import {Spinner} from "../commons/Spinner";
